test: split projection fixture helpers
